ntpull/ntpush - do not leave the .gpg file around when SIGTERMed.

+++ bin/ntcommon.sh
@@ -10,5 +10,12 @@ MONTHLY_ENCRYPTED_NOTES=${NOTES_DIR}/notes.${TIMESTAMP}.md.gpg
 ntpull() {
         rclone copy cloud:Notes/notes.md.gpg ${NOTES_DIR}
         gpg --yes -q -o ${PLAINTEXT_NOTES} -d ${ENCRYPTED_NOTES}
-        rm ${ENCRYPTED_NOTES}
 }
+
+cleanup() {
+        if [[ -f ${ENCRYPTED_NOTES} ]]; then
+                rm ${ENCRYPTED_NOTES}
+        fi
+}
+
+trap cleanup EXIT
